requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.1.0
dataclasses-json==0.5.7
deepl==1.14.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
langchain==0.0.108
markdownify==0.11.6
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.22.1
openai==0.27.2
packaging==23.0
Pillow==9.0.1
pydantic==1.10.6
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
six==1.16.0
soupsieve==2.4
SQLAlchemy==1.4.46
tenacity==8.2.2
torch==1.10.2
torchaudio==0.10.2
torchvision==0.11.3
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2
